[skrifa] replace double traversal optimization with explicit check
