+++ skrifa/src/color/traversal.rs
@@ -1,15 +1,16 @@
-use std::{cmp::Ordering, ops::Range};
+use std::ops::Range;
 
+use raw::ReadError;
 use read_fonts::{
-    tables::colr::{CompositeMode, Extend},
-    types::{BoundingBox, GlyphId, Point},
+    tables::colr::CompositeMode,
+    types::{BoundingBox, GlyphId},
 };
 
 use super::{
     instance::{
-        resolve_clip_box, resolve_paint, ColorStops, ColrInstance, ResolvedColorStop, ResolvedPaint,
+        resolve_clip_box, resolve_paint, ColrInstance, MaybeBrush, ResolvedColorStop, ResolvedPaint,
     },
-    Brush, ColorPainter, ColorStop, PaintCachedColorGlyph, PaintError, Transform,
+    Brush, ColorPainter, ColorStop, PaintCachedColorGlyph, PaintError,
 };
 
 use crate::decycler::{Decycler, DecyclerError};
@@ -70,86 +71,6 @@ impl From<ResolvedColorStop> for ColorStop {
     }
 }
 
-fn make_sorted_resolved_stops(
-    stops: &ColorStops,
-    instance: &ColrInstance,
-    out_stops: &mut ColorStopVec,
-) {
-    let color_stop_iter = stops.resolve(instance).map(|stop| stop.into());
-    out_stops.clear();
-    for stop in color_stop_iter {
-        out_stops.push(stop);
-    }
-    out_stops.sort_by(|a, b| a.offset.partial_cmp(&b.offset).unwrap_or(Ordering::Equal));
-}
-
-struct CollectFillGlyphPainter<'a> {
-    brush_transform: Option<Transform>,
-    glyph_id: GlyphId,
-    parent_painter: &'a mut dyn ColorPainter,
-    pub optimization_success: bool,
-}
-
-impl<'a> CollectFillGlyphPainter<'a> {
-    fn new(parent_painter: &'a mut dyn ColorPainter, glyph_id: GlyphId) -> Self {
-        Self {
-            brush_transform: None,
-            glyph_id,
-            parent_painter,
-            optimization_success: true,
-        }
-    }
-}
-
-impl ColorPainter for CollectFillGlyphPainter<'_> {
-    fn push_transform(&mut self, transform: Transform) {
-        if self.optimization_success {
-            match self.brush_transform {
-                None => {
-                    self.brush_transform = Some(transform);
-                }
-                Some(ref mut existing_transform) => {
-                    *existing_transform *= transform;
-                }
-            }
-        }
-    }
-
-    fn pop_transform(&mut self) {
-        // Since we only support fill and and transform operations, we need to
-        // ignore a popped transform, as this would be called after traversing
-        // the graph backup after a fill was performed, but we want to preserve
-        // the transform in order to be able to return it.
-    }
-
-    fn fill(&mut self, brush: Brush<'_>) {
-        if self.optimization_success {
-            self.parent_painter
-                .fill_glyph(self.glyph_id, self.brush_transform, brush);
-        }
-    }
-
-    fn push_clip_glyph(&mut self, _: GlyphId) {
-        self.optimization_success = false;
-    }
-
-    fn push_clip_box(&mut self, _: BoundingBox<f32>) {
-        self.optimization_success = false;
-    }
-
-    fn pop_clip(&mut self) {
-        self.optimization_success = false;
-    }
-
-    fn push_layer(&mut self, _: CompositeMode) {
-        self.optimization_success = false;
-    }
-
-    fn pop_layer(&mut self) {
-        self.optimization_success = false;
-    }
-}
-
 pub(crate) fn traverse_with_callbacks(
     paint: &ResolvedPaint,
     instance: &ColrInstance,
@@ -178,329 +99,62 @@ pub(crate) fn traverse_with_callbacks(
             }
             Ok(())
         }
-        ResolvedPaint::Solid {
-            palette_index,
-            alpha,
-        } => {
-            painter.fill(Brush::Solid {
-                palette_index: *palette_index,
-                alpha: *alpha,
-            });
-            Ok(())
-        }
-        ResolvedPaint::LinearGradient {
-            x0,
-            y0,
-            x1,
-            y1,
-            x2,
-            y2,
-            color_stops,
-            extend,
-        } => {
-            let mut p0 = Point::new(*x0, *y0);
-            let p1 = Point::new(*x1, *y1);
-            let p2 = Point::new(*x2, *y2);
-
-            let dot_product = |a: Point<f32>, b: Point<f32>| -> f32 { a.x * b.x + a.y * b.y };
-            let cross_product = |a: Point<f32>, b: Point<f32>| -> f32 { a.x * b.y - a.y * b.x };
-            let project_onto = |vector: Point<f32>, point: Point<f32>| -> Point<f32> {
-                let length = (point.x * point.x + point.y * point.y).sqrt();
-                if length == 0.0 {
-                    return Point::default();
-                }
-                let mut point_normalized = point / length;
-                point_normalized *= dot_product(vector, point) / length;
-                point_normalized
-            };
-
-            make_sorted_resolved_stops(color_stops, instance, resolved_stops);
-
-            // If p0p1 or p0p2 are degenerate probably nothing should be drawn.
-            // If p0p1 and p0p2 are parallel then one side is the first color and the other side is
-            // the last color, depending on the direction.
-            // For now, just use the first color.
-            if p1 == p0 || p2 == p0 || cross_product(p1 - p0, p2 - p0) == 0.0 {
-                if let Some(stop) = resolved_stops.first() {
-                    painter.fill(Brush::Solid {
-                        palette_index: stop.palette_index,
-                        alpha: stop.alpha,
-                    });
-                };
-                return Ok(());
-            }
-
-            // Follow implementation note in nanoemoji:
-            // https://github.com/googlefonts/nanoemoji/blob/0ac6e7bb4d8202db692574d8530a9b643f1b3b3c/src/nanoemoji/svg.py#L188
-            // to compute a new gradient end point P3 as the orthogonal
-            // projection of the vector from p0 to p1 onto a line perpendicular
-            // to line p0p2 and passing through p0.
-            let mut perpendicular_to_p2 = p2 - p0;
-            perpendicular_to_p2 = Point::new(perpendicular_to_p2.y, -perpendicular_to_p2.x);
-            let mut p3 = p0 + project_onto(p1 - p0, perpendicular_to_p2);
-
-            match (
-                resolved_stops.first().cloned(),
-                resolved_stops.last().cloned(),
-            ) {
-                (None, _) | (_, None) => {}
-                (Some(first_stop), Some(last_stop)) => {
-                    let mut color_stop_range = last_stop.offset - first_stop.offset;
-
-                    // Nothing can be drawn for this situation.
-                    if color_stop_range == 0.0 && extend != &Extend::Pad {
-                        return Ok(());
-                    }
-
-                    // In the Pad case, for providing normalized stops in the 0 to 1 range to the client,
-                    // insert a color stop at the end. Adding this stop will paint the equivalent gradient,
-                    // because: All font-specified color stops are in the same spot, mode is pad, so
-                    // everything before this spot is painted with the first color, everything after this spot
-                    // is painted with the last color. Not adding this stop would skip the projection below along
-                    // the p0-p3 axis and result in specifying non-normalized color stops to the shader.
-
-                    if color_stop_range == 0.0 && extend == &Extend::Pad {
-                        let mut extra_stop = last_stop;
-                        extra_stop.offset += 1.0;
-                        resolved_stops.push(extra_stop);
-
-                        color_stop_range = 1.0;
-                    }
-
-                    debug_assert!(color_stop_range != 0.0);
-
-                    if color_stop_range != 1.0 || first_stop.offset != 0.0 {
-                        let p0_p3 = p3 - p0;
-                        let p0_offset = p0_p3 * first_stop.offset;
-                        let p3_offset = p0_p3 * last_stop.offset;
-
-                        p3 = p0 + p3_offset;
-                        p0 += p0_offset;
-
-                        let scale_factor = 1.0 / color_stop_range;
-                        let start_offset = first_stop.offset;
-
-                        for stop in resolved_stops.iter_mut() {
-                            stop.offset = (stop.offset - start_offset) * scale_factor;
-                        }
-                    }
-
-                    painter.fill(Brush::LinearGradient {
-                        p0,
-                        p1: p3,
-                        color_stops: resolved_stops.as_slice(),
-                        extend: *extend,
-                    });
-                }
-            }
-
-            Ok(())
-        }
-        ResolvedPaint::RadialGradient {
-            x0,
-            y0,
-            radius0,
-            x1,
-            y1,
-            radius1,
-            color_stops,
-            extend,
-        } => {
-            let mut c0 = Point::new(*x0, *y0);
-            let mut c1 = Point::new(*x1, *y1);
-            let mut radius0 = *radius0;
-            let mut radius1 = *radius1;
-
-            make_sorted_resolved_stops(color_stops, instance, resolved_stops);
-
-            match (
-                resolved_stops.first().cloned(),
-                resolved_stops.last().cloned(),
-            ) {
-                (None, _) | (_, None) => {}
-                (Some(first_stop), Some(last_stop)) => {
-                    let mut color_stop_range = last_stop.offset - first_stop.offset;
-                    // Nothing can be drawn for this situation.
-                    if color_stop_range == 0.0 && extend != &Extend::Pad {
-                        return Ok(());
-                    }
-
-                    // In the Pad case, for providing normalized stops in the 0 to 1 range to the client,
-                    // insert a color stop at the end. See LinearGradient for more details.
-
-                    if color_stop_range == 0.0 && extend == &Extend::Pad {
-                        let mut extra_stop = last_stop;
-                        extra_stop.offset += 1.0;
-                        resolved_stops.push(extra_stop);
-                        color_stop_range = 1.0;
-                    }
-
-                    debug_assert!(color_stop_range != 0.0);
-
-                    // If the colorStopRange is 0 at this point, the default behavior of the shader is to
-                    // clamp to 1 color stops that are above 1, clamp to 0 for color stops that are below 0,
-                    // and repeat the outer color stops at 0 and 1 if the color stops are inside the
-                    // range. That will result in the correct rendering.
-                    if color_stop_range != 1.0 || first_stop.offset != 0.0 {
-                        let c0_to_c1 = c1 - c0;
-                        let radius_diff = radius1 - radius0;
-                        let scale_factor = 1.0 / color_stop_range;
-
-                        let c0_offset = c0_to_c1 * first_stop.offset;
-                        let c1_offset = c0_to_c1 * last_stop.offset;
-                        let stops_start_offset = first_stop.offset;
-
-                        // Order of reassignments is important to avoid shadowing variables.
-                        c1 = c0 + c1_offset;
-                        c0 += c0_offset;
-                        radius1 = radius0 + radius_diff * last_stop.offset;
-                        radius0 += radius_diff * first_stop.offset;
-
-                        for stop in resolved_stops.iter_mut() {
-                            stop.offset = (stop.offset - stops_start_offset) * scale_factor;
-                        }
-                    }
-
-                    painter.fill(Brush::RadialGradient {
-                        c0,
-                        r0: radius0,
-                        c1,
-                        r1: radius1,
-                        color_stops: resolved_stops.as_slice(),
-                        extend: *extend,
-                    });
-                }
+        ResolvedPaint::Solid { .. }
+        | ResolvedPaint::LinearGradient { .. }
+        | ResolvedPaint::RadialGradient { .. }
+        | ResolvedPaint::SweepGradient { .. } => {
+            if let MaybeBrush::Some(brush) = paint.as_brush(instance, resolved_stops)? {
+                painter.fill(brush);
             }
             Ok(())
         }
-        ResolvedPaint::SweepGradient {
-            center_x,
-            center_y,
-            start_angle,
-            end_angle,
-            color_stops,
-            extend,
-        } => {
-            // OpenType 1.9.1 adds a shift to the angle to ease specification of a 0 to 360
-            // degree sweep.
-            let sweep_angle_to_degrees = |angle| angle * 180.0 + 180.0;
-
-            let start_angle = sweep_angle_to_degrees(start_angle);
-            let end_angle = sweep_angle_to_degrees(end_angle);
-
-            // Stop normalization for sweep:
-
-            let sector_angle = end_angle - start_angle;
-
-            make_sorted_resolved_stops(color_stops, instance, resolved_stops);
-            if resolved_stops.is_empty() {
-                return Ok(());
-            }
-
-            match (
-                resolved_stops.first().cloned(),
-                resolved_stops.last().cloned(),
-            ) {
-                (None, _) | (_, None) => {}
-                (Some(first_stop), Some(last_stop)) => {
-                    let mut color_stop_range = last_stop.offset - first_stop.offset;
-
-                    let mut start_angle_scaled = start_angle + sector_angle * first_stop.offset;
-                    let mut end_angle_scaled = start_angle + sector_angle * last_stop.offset;
-
-                    let start_offset = first_stop.offset;
-
-                    // Nothing can be drawn for this situation.
-                    if color_stop_range == 0.0 && extend != &Extend::Pad {
-                        return Ok(());
-                    }
-
-                    // In the Pad case, if the color_stop_range is 0 insert a color stop at the end before
-                    // normalizing. Adding this stop will paint the equivalent gradient, because: All font
-                    // specified color stops are in the same spot, mode is pad, so everything before this
-                    // spot is painted with the first color, everything after this spot is painted with
-                    // the last color. Not adding this stop will skip the projection and result in
-                    // specifying non-normalized color stops to the shader.
-                    if color_stop_range == 0.0 && extend == &Extend::Pad {
-                        let mut offset_last = last_stop;
-                        offset_last.offset += 1.0;
-                        resolved_stops.push(offset_last);
-                        color_stop_range = 1.0;
-                    }
-
-                    debug_assert!(color_stop_range != 0.0);
-
-                    let scale_factor = 1.0 / color_stop_range;
-
-                    for shift_stop in resolved_stops.iter_mut() {
-                        shift_stop.offset = (shift_stop.offset - start_offset) * scale_factor;
-                    }
-
-                    // /* https://docs.microsoft.com/en-us/typography/opentype/spec/colr#sweep-gradients
-                    //  * "The angles are expressed in counter-clockwise degrees from
-                    //  * the direction of the positive x-axis on the design
-                    //  * grid. [...]  The color line progresses from the start angle
-                    //  * to the end angle in the counter-clockwise direction;" -
-                    //  * Convert angles and stops from counter-clockwise to clockwise
-                    //  * for the shader if the gradient is not already reversed due to
-                    //  * start angle being larger than end angle. */
-                    start_angle_scaled = 360.0 - start_angle_scaled;
-                    end_angle_scaled = 360.0 - end_angle_scaled;
-
-                    if start_angle_scaled >= end_angle_scaled {
-                        (start_angle_scaled, end_angle_scaled) =
-                            (end_angle_scaled, start_angle_scaled);
-                        resolved_stops.reverse();
-                        for stop in resolved_stops.iter_mut() {
-                            stop.offset = 1.0 - stop.offset;
-                        }
-                    }
-
-                    // https://learn.microsoft.com/en-us/typography/opentype/spec/colr#sweep-gradients
-                    // "If the color line's extend mode is reflect or repeat
-                    // and start and end angle are equal, nothing shall be drawn."
-                    if start_angle_scaled == end_angle_scaled && extend != &Extend::Pad {
-                        return Ok(());
-                    }
-
-                    painter.fill(Brush::SweepGradient {
-                        c0: Point::new(*center_x, *center_y),
-                        start_angle: start_angle_scaled,
-                        end_angle: end_angle_scaled,
-                        color_stops: resolved_stops.as_slice(),
-                        extend: *extend,
-                    });
-                }
-            }
-            Ok(())
-        }
-
         ResolvedPaint::Glyph { glyph_id, paint } => {
             let glyph_id = (*glyph_id).into();
-            let mut optimizer = CollectFillGlyphPainter::new(painter, glyph_id);
-            let mut result = traverse_with_callbacks(
-                &resolve_paint(instance, paint)?,
+            // Look for the pattern `(transform)* fill` and optimize it to a
+            // single paint call
+            let mut next_paint = resolve_paint(instance, paint)?;
+            // Collect any chain of intermediate transforms
+            let mut intermediate_transform = None;
+            while let Some((transform, child_paint)) = next_paint.as_transform() {
+                intermediate_transform = Some(match intermediate_transform {
+                    Some(existing_transform) => existing_transform * transform,
+                    None => transform,
+                });
+                next_paint = resolve_paint(instance, &child_paint)?;
+            }
+            // If the next paint is a brush, we can optimize the traversal to a
+            // single fill_glyph call
+            match next_paint.as_brush(instance, resolved_stops)? {
+                MaybeBrush::Some(brush) => {
+                    painter.fill_glyph(glyph_id, intermediate_transform, brush);
+                    return Ok(());
+                }
+                // Valid brush but doesn't produce any rendering, so we can
+                // skip the glyph entirely
+                MaybeBrush::NonRendering => {
+                    return Ok(());
+                }
+                // Not a brush, fall through to the unoptimized path
+                MaybeBrush::None => {}
+            }
+            // In case the optimization was not successful, just push a clip,
+            // and continue unoptimized traversal.
+            painter.push_clip_glyph(glyph_id);
+            if let Some(transform) = intermediate_transform {
+                painter.push_transform(transform);
+            }
+            let result = traverse_with_callbacks(
+                &next_paint,
                 instance,
-                &mut optimizer,
+                painter,
                 decycler,
                 resolved_stops,
                 recurse_depth + 1,
             );
-
-            // In case the optimization was not successful, just push a clip, and continue unoptimized traversal.
-            if !optimizer.optimization_success {
-                painter.push_clip_glyph(glyph_id);
-                result = traverse_with_callbacks(
-                    &resolve_paint(instance, paint)?,
-                    instance,
-                    painter,
-                    decycler,
-                    resolved_stops,
-                    recurse_depth + 1,
-                );
-                painter.pop_clip();
+            if intermediate_transform.is_some() {
+                painter.pop_transform();
             }
-
+            painter.pop_clip();
             result
         }
         ResolvedPaint::ColrGlyph { glyph_id } => {
@@ -551,7 +205,12 @@ pub(crate) fn traverse_with_callbacks(
         | ResolvedPaint::Skew {
             paint: next_paint, ..
         } => {
-            painter.push_transform(paint.try_into()?);
+            painter.push_transform(
+                paint
+                    .as_transform()
+                    .ok_or(ReadError::MalformedData("expected a transform paint"))?
+                    .0,
+            );
             let result = traverse_with_callbacks(
                 &resolve_paint(instance, next_paint)?,
                 instance,
